Change the way new certs are detected.
Previously, we relied upon mtime to detect if we needed to install
new certs.  This was wrong. The original use of fetch wanted to
exploit the -m (Mirror mode) option so that only newly downloaded
certs were later installed.

This would be the ideal situation:

  * Certs are renewly daily

  * Depending on crontabs on each client, it might take up to 48 hours
    for that cert to be downloaded onto the server on which it would be
    used. This allows for propogation from the cert server to the rsync
    jail to the web server we download from. Sure, we can get this down
    to an hour or so, but things go wrong. This client must be resilient.

  * A new cert shows up, we see a recent mtime, we install it.

However, this leads to some difficult edge cases:

  * Certs where created 10 days go, we're just setting up, mtime
    fails

  * The cert propagation chain fails for a while, when fixed, we've
    missed cert changes because they are too old.